Move logic to action button component
