+++ projects/packages/my-jetpack/_inc/components/connected-product-card/index.tsx
@@ -46,8 +46,7 @@ const ConnectedProductCard: FC< ConnectedProductCardProps > = ( {
 	customLoadTracks,
 } ) => {
 	const { isRegistered } = useMyJetpackConnection();
-
-	const { detail, refetch, isLoading: isProductDataLoading } = useProduct( slug );
+	const { detail, refetch } = useProduct( slug );
 	const { name, description: defaultDescription, status, manageUrl } = detail;
 
 	const DefaultDescription = () => {
@@ -74,7 +73,6 @@ const ConnectedProductCard: FC< ConnectedProductCardProps > = ( {
 			status={ status }
 			admin={ admin }
 			recommendation={ recommendation }
-			isFetching={ isProductDataLoading }
 			isDataLoading={ isDataLoading }
 			additionalActions={ additionalActions }
 			primaryActionOverride={ primaryActionOverride }
